reddit_clone/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-axes==6.1.1
django-ratelimit==4.1.0
channels==4.0.0
channels-redis==4.1.0
django-guardian==2.4.0
drf-yasg==1.21.7
argon2-cffi==23.1.0
bleach==6.1.0
psycopg2-binary==2.9.9
redis==5.0.1
daphne==4.0.0
markdown==3.5.1
python-magic==0.4.27
pyotp==2.9.0
qrcode==7.4.2

# For development
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.23.2
coverage==7.3.4
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0 